Removed  warnings

+++ modules/diffNaive.h
@@ -25,7 +25,7 @@ Change diffNaive(const std::string& student, const std::string& instructor){
 	differences.a_start = 0;
 	differences.b_start = 0;
 
-	unsigned int len = (student.size() < instructor.size()) ? student.length() : instructor.length();
+	unsigned int len = (unsigned int)(student.size() < instructor.size()) ? (unsigned int)student.length() : (unsigned int)instructor.length();
 
 	unsigned int i = 0;
 	for(i = 0; i < len; i++){
@@ -125,7 +125,7 @@ Difference diffLine(const std::string& _student, const std::string& _instructor)
 		}
 		i++;
 	}
-	diffs.distance = std::max(file.a_changes.size(), file.b_changes.size());
+	diffs.distance = (int)std::max(file.a_changes.size(), file.b_changes.size());
 	diffs.changes.push_back(file);
 	return diffs;
 }
@@ -170,7 +170,7 @@ Difference diffLineNoSpace(const std::string& _student, const std::string& _inst
 		}
 		i++;
 	}
-	diffs.distance = std::max(file.a_changes.size(), file.b_changes.size());
+	diffs.distance = (int)std::max(file.a_changes.size(), file.b_changes.size());
 	diffs.changes.push_back(file);
 	return diffs;
 }
